requirements.txt
opencv-python
numpy
kaggle
python-dotenv
torch
torchvision
selenium
replicate